fixing tags on client

+++ deploy/deploy-media.xfer.sh
@@ -9,6 +9,11 @@ fi
 
 root=$(dirname $0)/..
 
+if [ -f $root/media/library/library.json ]; then
+  echo "Refusing to overwrite library.json";
+  exit 1;
+fi
+
 cp --no-dereference --recursive $root/media/* $TARGET_DIR
 cp $root/deploy/xfer-media.config.js $TARGET_DIR/server/config.override.js
 cp $root/deploy/rescan.cron.sh $TARGET_DIR/rescan.cron.sh
